pyproject.toml
[project]
name = "delta-digest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "delta-spark==3.2.1",
    "feedparser>=6.0.12",
    "google-genai>=1.66.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pydantic-settings>=2.13.1",
    "pyspark==3.5.4",
    "pyyaml>=6.0.3",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = [
    "hatchling>=1.29.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
